% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui_assessmentOutput.R
\name{assessmentOutput}
\alias{assessmentOutput}
\title{Create assessment output}
\usage{
assessmentOutput(pool = NULL, config = NULL, overwrite = F)
}
\arguments{
\item{pool}{Pool}

\item{config}{Config}

\item{overwrite}{T/F}
}
\value{
code to be used as argument to the function 'shinyApp(ui=...)'
}
\description{
Function to be passed as argument to 'shinyApp(ui=...)'
}
\examples{
# default
shinyApp(assessmentOutput(pool = getDemoPool("demo01"), config = getConfig(), overwrite=T), renderAssessment)
}
