% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_from_ib_file.R
\name{get_metadata_from_ib_file}
\alias{get_metadata_from_ib_file}
\title{Read Metadata in CBA ItemBuilder Project Files}
\usage{
get_metadata_from_ib_file(zip_path, metadata_xml_filename = "metadata.xml")
}
\arguments{
\item{zip_path}{Reference to the CBA ItemBuilder project file (relative path of the ZIP archive including the file name)}

\item{metadata_xml_filename}{Internal file name in CBA ItemBuilder projects (should be "metadata.xml")}
}
\value{
data frame with all defined metadata
}
\description{
`get_metadata_from_ib_file` inspects a single CBA ItemBuilder project file
}
\details{
The returned object is a data.frame containing the metadata information defined in a CBA ItemBuilder project file
}
\examples{
\dontrun{
 demo <- get_metadata_from_ib_file("ib/myproject.zip")
}
}
