% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_properties_from_ib_file.R
\name{get_properties_from_ib_file}
\alias{get_properties_from_ib_file}
\title{Extract and Parse Property Information from CBA ItemBuilder Project Files}
\usage{
get_properties_from_ib_file(
  zip_path,
  properties_filename = "project.properties"
)
}
\arguments{
\item{zip_path}{Reference to the CBA ItemBuilder project file (relative path of the ZIP archive including the file name)}

\item{properties_filename}{Internal file name in CBA ItemBuilder projects (should be "project.properties")}
}
\value{
data frame with property information
}
\description{
`get_properties_from_ib_file` inspects a single CBA ItemBuilder project file
}
\details{
The returned object is a data.frame containing properties defined in a CBA ItemBuilder project file
}
\examples{
\dontrun{
 demo <- get_properties_from_ib_file("ib/myproject.zip")
}
}
