% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stimulus_infos_from_ib_file.R
\name{get_stimulus_infos_from_ib_file}
\alias{get_stimulus_infos_from_ib_file}
\title{Read Stimuls Definition from CBA ItemBuilder Project File}
\usage{
get_stimulus_infos_from_ib_file(
  zip_path,
  scoring_json_filename = "stimulus.json"
)
}
\arguments{
\item{zip_path}{Reference to the CBA ItemBuilder project file (relative path of the ZIP archive including the file name)}

\item{scoring_json_filename}{Internal file name in CBA ItemBuilder projects (should be "stimulus.json")}
}
\value{
data frame with stimulus information
}
\description{
`get_stimulus_infos_from_ib_file` inspects a single CBA ItemBuilder project file
}
\details{
The returned object is a data.frame containing the stimulus information defined in a CBA ItemBuilder project file
}
\examples{
\dontrun{
 demo <- get_stimulus_infos_from_ib_file("ib/myproject.zip")
}
}
