% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value_maps_from_ib_file.R
\name{get_value_maps_from_ib_file}
\alias{get_value_maps_from_ib_file}
\title{Extract and Parse Value Map Information from CBA ItemBuilder Project Files}
\usage{
get_value_maps_from_ib_file(
  zip_path,
  valuemap_xml_filename = "global.cbavaluemap"
)
}
\arguments{
\item{zip_path}{Reference to the CBA ItemBuilder project file (relative path of the ZIP archive including the file name)}

\item{valuemap_xml_filename}{Internal file name in CBA ItemBuilder projects (should be "global.cbavaluemap")}
}
\value{
data frame with defined value maps
}
\description{
`get_value_maps_from_ib_file` inspects a single CBA ItemBuilder project file
}
\details{
The returned object is a data.frame containing information for value maps defined in a CBA ItemBuilder project file
}
\examples{
\dontrun{
 demo <- get_value_maps_from_ib_file("ib/myproject.zip")
}
}
