% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_infos_from_ib_file.R
\name{get_variable_infos_from_ib_file}
\alias{get_variable_infos_from_ib_file}
\title{Extract and Parse Variable Information from CBA ItemBuilder Project Files}
\usage{
get_variable_infos_from_ib_file(
  zip_path,
  stimulus_json_filename = "stimulus.json",
  variables_xml_file = "global.cbavariables"
)
}
\arguments{
\item{zip_path}{Reference to the CBA ItemBuilder project file (relative path of the ZIP archive including the file name)}

\item{stimulus_json_filename}{Internal file name in CBA ItemBuilder projects (should be "stimulus.json")}

\item{variables_xml_file}{Internal file name in CBA ItemBuilder projects (should be "global.cbavariables")}
}
\value{
data frame with property information
}
\description{
`get_variable_infos_from_ib_file` inspects a single CBA ItemBuilder project file
}
\details{
The returned object is a data.frame containing variables defined in a CBA ItemBuilder project file
}
\examples{
\dontrun{
 demo <- get_variable_infos_from_ib_file("ib/myproject.zip")
}
}
