% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{getConfig}
\alias{getConfig}
\title{Get configuration object}
\usage{
getConfig(
  WindowTitle = "MyAssessment",
  Verbose = FALSE,
  WWWfolder = "_mywww",
  Datafolder = "_mydata",
  posH = "center",
  posV = "center",
  scaling = "scale-up-down",
  sessiontype = "sessionstorage",
  maintenancePassword = "",
  maintenanceQuery = "maintenance",
  maintenanceKey = list(key = "x", ctrl = T, shift = F, alt = F)
)
}
\arguments{
\item{WindowTitle}{Title of the html browser window.}

\item{Verbose}{Should the package provide log information to the console?}

\item{WWWfolder}{Folder to store html/javascript files required to run the assessment}

\item{Datafolder}{Folder to store data files for started/completed assessments}

\item{posH}{Horizontal orientation (should be one of 'left', 'right' or 'center')}

\item{posV}{Vertical orientation (should be one of 'top', 'bottom' or 'center')}

\item{scaling}{Scaling of content (should be one of 'scale-up', 'scale-down', 'scale-up-down' or 'no-scaling')}

\item{sessiontype}{Session storage (should be one of 'sessionstorage', 'cookie', 'localstorage' or 'provided' )}

\item{maintenancePassword}{Password to access data online (no access possible if not defined).}

\item{maintenanceQuery}{Query string parameter name to access data online.}

\item{maintenanceKey}{Keyboard shortcut to access the maintenance dialog (provide a list in the following form: `list(key="X", ctrl=T, shift=F, alt=F)`)}
}
\value{
config object (list)
}
\description{
`getConfig` returns a config object required to use ShinyItemBuilder.
}
\details{
The returned object is a list that contains the different configurations  for an assessment using ShinyItemBuilder.
}
\examples{
\dontrun{
 conf <- getConfig()
}
}
