% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pool.R
\name{getDemoPool}
\alias{getDemoPool}
\title{Get example item pool}
\usage{
getDemoPool(name = "demo01")
}
\arguments{
\item{name}{Name of the example item pool (currently: "demo01").}
}
\value{
config object (list)
}
\description{
Get example item pool for demonstration purposes using items included in the ShinyItemBuilder package for demonstration purposes
}
\examples{
# default
pool <- getDemoPool("demo01")
}
