% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pool.R
\name{getPool}
\alias{getPool}
\title{Get item pool}
\usage{
getPool(path = NULL, files = NULL, tasks = NULL, scope = NULL)
}
\arguments{
\item{path}{Path to a folder with CBA ItemBuilder project files.}

\item{files}{List of files (i.e., path and file names) of CBA ItemBuilder project files.}

\item{tasks}{List of tasks (need to have the same length as files).}

\item{scope}{Scope (either single string or list of the same length as files / tasks).}
}
\value{
config object (list)
}
\description{
Get item pool for an assessment with ShinyItemBuilder
}
\examples{
# default
pool <- getPool(path="../")
}
