% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{getValueForTestTaker}
\alias{getValueForTestTaker}
\title{Get (i.e., load / retrieve) a value for a particular test-taker}
\usage{
getValueForTestTaker(session, name, default = NULL, store = T)
}
\arguments{
\item{session}{The shiny session object.}

\item{name}{Name of the value.}

\item{default}{Default value.}

\item{store}{Should the value be stored, if the default is used?}
}
\value{
The value (if present) for the current test-taker (or the default value)
}
\description{
Get (i.e., load / retrieve) a value for a particular test-taker
}
\examples{
getValueForTestTaker(session, "FirstItem",1)
getValueForTestTaker(session, "FirstItem",1, T)
}
