% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_project_tool.R
\name{inspect_project_tool}
\alias{inspect_project_tool}
\title{inspect_project_tool.R}
\usage{
inspect_project_tool()
}
\description{
This function starts the "inspect-project-tool" (ipt, a shiny app included in ShinyItemBuilder)
}
