% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_ib_trace.R
\name{parse_ib_trace}
\alias{parse_ib_trace}
\title{Parse CBA ItemBuilder Trace JSON}
\usage{
parse_ib_trace(tracejson)
}
\arguments{
\item{tracejson}{Trace JSON as provided by the CBA ItemBuilder runtime.}
}
\value{
data frame with all hits and result texts
}
\description{
`parse_ib_trace` converts the JSON to a data.frame
}
\details{
The returned object returns a data.frame containing the trace data provided as JSON.
}
\examples{
\dontrun{
 demo <- parse_ib_trace(json)
}
}
