% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_renderAssessment.R
\name{renderAssessment}
\alias{renderAssessment}
\title{Render assessment content}
\usage{
renderAssessment(input, output, session)
}
\arguments{
\item{input}{Input.}

\item{output}{Output}

\item{session}{Session.}
}
\value{
code to be used as argument to the function 'shinyApp(server=...)'
}
\description{
Function to be passed as argument to 'shinyApp(server=...)'
}
\examples{
\dontrun{
shinyApp(assessmentOutput(pool = getDemoPool("demo01"),
     config = getConfig(), overwrite=T),
     renderAssessment)
}
}
