% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scoring_info_from_ib_file.R
\name{get_scoring_info_from_ib_file}
\alias{get_scoring_info_from_ib_file}
\title{Parse Scoring Information in CBA ItemBuilder Project Files}
\usage{
get_scoring_info_from_ib_file(
  zip_path,
  scoring_xml_filename = "global.cbaitemscore"
)
}
\arguments{
\item{zip_path}{Reference to the CBA ItemBuilder project file (relative path of the ZIP archive including the file name)}

\item{scoring_xml_filename}{Internal file name in CBA ItemBuilder projects (should be "global.cbaitemscore")}
}
\value{
data frame with all classes and hits
}
\description{
`get_scoring_info_from_ib_file` inspects a single CBA ItemBuilder project file
}
\details{
The returned object returns a data.frame containing the scoring information defined in a CBA ItemBuilder project file
}
\examples{
\dontrun{
 demo <- get_scoring_info_from_ib_file("ib/myproject.zip")
}
}
