% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config.R
\name{getConfig}
\alias{getConfig}
\title{Get configuration object (currently a list) with examples}
\usage{
getConfig(
  WindowTitle = "MyAssessment",
  Verbose = TRUE,
  WWWfolder = "_mywww",
  Datafolder = "_mydata",
  posH = "center",
  posV = "center",
  scaling = "updown",
  sessiontype = "sessionstorage",
  maintenancePassword = ""
)
}
\arguments{
\item{WindowTitle}{Title of the html browser window.}

\item{Verbose}{Should the package provide log information to the console?}

\item{WWWfolder}{Folder to store html/javascript files required to run the assessment}

\item{Datafolder}{Folder to store data files for started/completed assessments}

\item{posH}{Horizontal orientation (should be one of 'left', 'right' or 'center')}

\item{posV}{Vertical orientation (should be one of 'top', 'bottom' or 'center')}

\item{scaling}{Scaling of content (should be one of 'up', 'down', 'updown' or 'none')}

\item{sessiontype}{Session storage (should be one of 'sessionstorage', 'cookie', 'localstorage' or 'provided' )}

\item{maintenancePassword}{Password to access data online (no access possible if not defined).}
}
\value{
The value (if present) for the current test-taker (or the default value)
}
\description{
Get configuration object (currently a list) with examples
}
