% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_storage.R
\name{setValueForTestTaker}
\alias{setValueForTestTaker}
\title{Set (i.e., save, store) a value for a particular test-taker}
\usage{
setValueForTestTaker(session, name, value)
}
\arguments{
\item{session}{The shiny session object.}

\item{name}{Name of the value.}

\item{value}{Value to store.}
}
\value{
The value (if present) for the current test-taker (or the default value)
}
\description{
`getValueForTestTaker` can be used in a custom navigation
             function to store data for the current session.
}
\examples{
# default
setValueForTestTaker(session, "current-item",1)
}
